requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.10.4
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
nodeenv==1.9.1
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
pillow==11.0.0
prisma==0.15.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
StrEnum==0.4.15
tomlkit==0.13.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
"python-telegram-bot[job-queue]"
